Revert "NodeJS Compat: process features (#4271)"
This reverts commit 3edffd2aa4ef8c151df1a334236fa8668d8b396b.
