+++ src/node/events.ts
@@ -3,20 +3,5 @@
 //     https://opensource.org/licenses/Apache-2.0
 //
 
-// Explicit list used to hide _initialized private export
-export {
-  EventEmitter,
-  EventEmitterAsyncResource,
-  addAbortListener,
-  captureRejectionSymbol,
-  errorMonitor,
-  getMaxListeners,
-  usingDomains,
-  defaultMaxListeners,
-  setMaxListeners,
-  listenerCount,
-  getEventListeners,
-  once,
-  on,
-} from 'node-internal:events';
+export * from 'node-internal:events';
 export { default } from 'node-internal:events';
