+++ src/node/internal/events.ts
@@ -42,7 +42,7 @@ import {
   validateObject,
 } from 'node-internal:validators';
 
-import * as process from 'node-internal:internal_process';
+import * as process from 'node-internal:process';
 
 import { spliceOne } from 'node-internal:internal_utils';
 
@@ -163,8 +163,6 @@ export function addAbortListener(signal: AbortSignal, listener: any) {
 }
 
 export default EventEmitter;
-
-// NOTE: all properties of event emitter must also be re-exported by the events.ts wrapper
 EventEmitter.on = on;
 EventEmitter.once = once;
 EventEmitter.getEventListeners = getEventListeners;
@@ -497,32 +495,24 @@ function _addListener(
           `added to an EventEmitter. Use ` +
           'emitter.setMaxListeners() to increase limit'
       );
-      const w: EventEmitterError = Object.assign(
-        new Error(
-          'Possible EventEmitter memory leak detected. ' +
-            `${existing.length} ${String(type)} listeners ` +
-            `added to ${inspect(target, { depth: -1 })}. Use ` +
-            'emitter.setMaxListeners() to increase limit'
-        ),
-        {
-          name: 'MaxListenersExceededWarning',
-          emitter: target,
-          type: type,
-          count: existing.length,
-        }
-      );
-      process.emitWarning(w);
+      // TODO(soon): Implement process.emitWarning and inspect
+      // // No error code for this since it is a Warning
+      // // eslint-disable-next-line no-restricted-syntax
+      // const w = new Error(
+      //   "Possible EventEmitter memory leak detected. " +
+      //     `${existing.length} ${String(type)} listeners ` +
+      //     `added to ${inspect(target, { depth: -1 })}. Use ` +
+      //     "emitter.setMaxListeners() to increase limit",
+      // );
+      // w.name = "MaxListenersExceededWarning";
+      // w.emitter = target;
+      // w.type = type;
+      // w.count = existing.length;
+      // process.emitWarning(w);
     }
   }
 
   return target;
-}
-
-interface EventEmitterError extends Error {
-  name: string;
-  emitter: unknown;
-  type: string | symbol;
-  count: number;
 }
 
 EventEmitter.prototype.addListener = function addListener(
@@ -1030,8 +1020,3 @@ export function on(
     iterator.return();
   }
 }
-
-// Initialize the process global as an event emitter, if process executed first
-if (process._initialized) process._initProcess();
-
-export var _initialized = true;
