+++ src/node/internal/internal_utils.ts
@@ -28,7 +28,7 @@
 
 import { default as bufferUtil } from 'node-internal:buffer';
 import type { Encoding } from 'node-internal:buffer';
-import { validateFunction, validateString } from 'node-internal:validators';
+import { validateFunction } from 'node-internal:validators';
 import {
   ERR_FALSY_VALUE_REJECTION,
   Falsy,
@@ -372,91 +372,4 @@ export function callbackify<T extends (...args: unknown[]) => Promise<unknown>>(
   // eslint-disable-next-line @typescript-eslint/ban-ts-comment
   // @ts-expect-error
   return callbackified;
-}
-
-export function parseEnv(content: string): Record<string, string> {
-  validateString(content, 'content');
-
-  const result: Record<string, string> = {};
-  const lines = content.split('\n');
-
-  for (let i = 0; i < lines.length; i++) {
-    let line = lines[i];
-    if (line === undefined) continue;
-
-    if (!line.trim()) continue;
-    if (line.trimStart().startsWith('#')) continue;
-    if (line.trimStart().startsWith('export '))
-      line = line.substring(line.indexOf('export ') + 7);
-
-    const equalIndex = line.indexOf('=');
-    if (equalIndex === -1) continue;
-
-    const key = line.substring(0, equalIndex).trim();
-    if (!key) continue;
-
-    let value = line.substring(equalIndex + 1).trimStart();
-    if (value.length > 0) {
-      const maybeQuote = value[0];
-      if (maybeQuote === '"' || maybeQuote === "'" || maybeQuote === '`') {
-        // Check if the closing quote is on the same line
-        const closeIndex = value.indexOf(maybeQuote, 1);
-
-        if (closeIndex !== -1) {
-          // Found closing quote on same line
-          value = value.substring(1, closeIndex);
-          // Only handle escape sequences for double quotes
-          if (maybeQuote === '"') value = value.replace(/\\n/g, '\n');
-          // For single quotes and backticks, keep \n as literal
-        } else {
-          // Check for multiline strings
-          let fullValue = value.substring(1); // Remove opening quote
-          let currentLine = i;
-          let foundClosingQuote = false;
-
-          // Look for closing quote in subsequent lines
-          while (currentLine < lines.length - 1) {
-            currentLine++;
-            const nextLine = lines[currentLine];
-            if (nextLine !== undefined) {
-              const closeInNextLine = nextLine.indexOf(maybeQuote);
-              if (closeInNextLine !== -1) {
-                // Found closing quote
-                fullValue += '\n' + nextLine.substring(0, closeInNextLine);
-                value = fullValue;
-
-                // Only handle escape sequences for double quotes
-                if (maybeQuote === '"') {
-                  value = value.replace(/\\n/g, '\n');
-                }
-
-                foundClosingQuote = true;
-                i = currentLine; // Update line counter
-                break;
-              } else {
-                // Continue building multiline value
-                fullValue += '\n' + nextLine;
-              }
-            }
-          }
-
-          if (!foundClosingQuote) {
-            if (value.length === 1) {
-              // Just the quote character, return it as the value
-              value = maybeQuote;
-            } else {
-              // Return content after the opening quote
-              value = value.substring(1);
-            }
-          }
-        }
-      } else {
-        const hashIndex = value.indexOf('#');
-        if (hashIndex !== -1) value = value.substring(0, hashIndex);
-        value = value.trimEnd();
-      }
-    }
-    result[key] = value;
-  }
-  return result;
 }
