--- src/node/internal/process.d.ts
@@ -1,7 +1,0 @@
-export const versions: Record<string, string>;
-export function getEnvObject(): Record<string, string>;
-export function getBuiltinModule(id: string): object;
-export function processExitImpl(code: number): void;
-export const processPlatform: string;
-export const nodeVersion: string;
-export const workerdVersion: string;
