+++ src/workerd/api/node/BUILD.bazel
@@ -10,7 +10,6 @@ wd_cc_library(
         "crypto-keys.c++",
         "diagnostics-channel.c++",
         "module.c++",
-        "process.c++",
         "timers.c++",
         "util.c++",
         "zlib-util.c++",
@@ -20,7 +19,6 @@ wd_cc_library(
         "diagnostics-channel.h",
         "module.h",
         "node.h",
-        "process.h",
         "timers.h",
         "util.h",
         "zlib-util.h",
@@ -54,7 +52,6 @@ wd_cc_library(
         "buffer-string-search.h",
         "dns.h",
         "i18n.h",
-        "node-version.h",
         "url.h",
     ],
     implementation_deps = [
