+++ src/workerd/api/node/tests/process-nodejs-test.wd-test
@@ -11,8 +11,7 @@ const unitTests :Workerd.Config = (
         compatibilityDate = "2024-12-28",
         compatibilityFlags = [
           "nodejs_compat",
-          "nodejs_compat_populate_process_env",
-          "experimental"
+          "nodejs_compat_populate_process_env"
         ],
         bindings = [
           (name = "FOO", text = "BAR"),
